refactor(operations): relocate protocols/structure to operations/schema
Move lionagi/protocols/structure/{base,json_structure}.py to
lionagi/operations/schema/{structure,json_structure}.py.

Structure and JsonStructure are LLM-output parsing utilities, not data
primitives. All 5 callers already live in operations/ or protocols/messages/:
  - operations/types.py
  - operations/parse/parse.py
  - operations/communicate/communicate.py  (lazy import)
  - operations/run/run.py                  (lazy import)
  - protocols/messages/instruction.py      (lazy import)

All 5 import sites updated to point to the new canonical location.

Backward-compat shims left at the old paths so any code importing
lionagi.protocols.structure.base, .json_structure, or the package
directly continues to work without changes (aliases, not copies).

Test at tests/protocols/messages/test_utils.py updated to use new path.
New test module tests/operations/schema/test_structure.py added with 17
tests covering Structure, JsonStructure, and the deprecation shims.
