+++ lionagi/operations/schema/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2023-2025, HaiyangLi <quantocean.li at gmail dot com>
+# SPDX-License-Identifier: Apache-2.0
+
+from .json_structure import JsonStructure
+from .structure import Structure
+
+__all__ = ("JsonStructure", "Structure")
